railway-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.0.1+cpu
transformers==4.35.0
numpy==1.24.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78